requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
gevent==21.8.0
greenlet==1.1.2
grequests==0.6.0
idna==3.2
lxml==4.6.3
Pillow==8.3.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.7
zope.event==4.5.0
zope.interface==5.4.0
